+++ pkg/docker/queries.go
@@ -1,4 +1,4 @@
-package utils
+package docker
 
 import "github.com/docker/docker/api/types/container"
